+++ test/rowcol.js
@@ -293,57 +293,4 @@ describe('rowcol', () => {
       });
     });
   });
-
-  describe('#proxy', () => {
-    /* Proxy Test */
-    let prox = rowcol.proxy(colData, 1);
-
-    it('should not commit when the value is set', () => {
-      prox.approved = false;
-      assert.equal(colData.approved[1], true, 'proxy: commit prior value');
-    });
-
-    it('should update the value when commit is called', () => {
-      prox.commit();
-      assert.equal(colData.approved[1], false, 'proxy: commit value mismatch');
-    });
-
-    it('should not commit when the value is set', () => {
-      prox.name = 'D';
-      assert.equal(colData.name[1], 'B', 'proxy: finalize prior value');
-    });
-
-    it('should update the value when finalize is called', () => {
-      prox.finalize();
-      assert.equal(colData.name[1], 'D', 'proxy: finalize value mismatch');
-    });
-
-    it('should not permit changes after finalize', () => {
-      assert.throws(() => {
-        prox.commit();
-      }, ReferenceError);
-    });
-
-    it('should support clearUndef', () => {
-      prox = rowcol.proxy(colDataWithUndefined, 1, true);
-
-      assert.equal('name' in prox, false);
-
-      // Add the name to the proxy
-      prox.name = 'B';
-      prox.commit();
-
-      assert.equal(colDataWithUndefined.name[1], 'B');
-
-      // clearUndef will not move undefineds back into the object
-      prox.name = undefined;
-      prox.finalize();
-
-      assert.equal(colDataWithUndefined.name[1], 'B');
-    });
-  });
-
-  describe('#rotate', () => {
-
-  });
 });
